[DP-47] refactor: purchaseTabs 컴포넌트 구조화 #47
