+++ src/components/common/tabs/PurchaseTabs.tsx
@@ -1,39 +1,25 @@
 "use client";
 
-import { useEffect, useState, useCallback, useRef } from "react";
+import { useEffect, useState } from "react";
+import { useMultiRef } from "./useMultiRef";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
+import { PURCHASE_MODE_TABS } from "./tabsConfig";
 import { motion } from "framer-motion";
 
 import { cn } from "@/lib/utils";
 
-interface PurchaseModeTabsProps {
+interface Props {
   value: string;
   onChange: (value: string) => void;
 }
 
-const TABS = [
-  { label: "일반 구매", value: "normal", content: "일반 구매 내용입니다." },
-  { label: "자투리 구매", value: "fragment", content: "자투리 구매 내용입니다." },
-];
-
-function useMultiRef<T>() {
-  const refs = useRef<(T | null)[]>([]);
-  const getRef = useCallback(
-    (i: number) => (el: T | null) => {
-      refs.current[i] = el;
-    },
-    []
-  );
-  return [refs, getRef] as const;
-}
-
-export default function PurchaseModeTabs({ value, onChange }: PurchaseModeTabsProps) {
+export default function PurchaseModeTabs({ value, onChange }: Props) {
   const [triggerRefs, setTriggerRef] = useMultiRef<HTMLButtonElement>();
   const [pendingValue, setPendingValue] = useState(value);
   const [highlightStyle, setHighlightStyle] = useState({ left: 0, width: 0 });
 
   useEffect(() => {
-    const index = TABS.findIndex((tab) => tab.value === pendingValue);
+    const index = PURCHASE_MODE_TABS.findIndex((tab) => tab.value === pendingValue);
     const current = triggerRefs.current[index];
     if (current) {
       const { offsetLeft, offsetWidth } = current;
@@ -57,7 +43,7 @@ export default function PurchaseModeTabs({ value, onChange }: PurchaseModeTabsPr
           animate={{ left: highlightStyle.left, width: highlightStyle.width }}
           className="absolute top-1 bottom-1 rounded-full bg-color-bg-secondary2 z-0"
         />
-        {TABS.map((tab, i) => (
+        {PURCHASE_MODE_TABS.map((tab, i) => (
           <TabsTrigger
             key={tab.value}
             value={tab.value}
@@ -65,7 +51,8 @@ export default function PurchaseModeTabs({ value, onChange }: PurchaseModeTabsPr
             className={cn(
               "relative z-10 flex-1 py-3 px-6 rounded-full text-sm font-bold transition-colors",
               "text-color-gray-600 data-[state=active]:text-color-text-black",
-              "data-[state=active]:bg-color-bg-secondary2"
+              "data-[state=active]:bg-color-bg-secondary2",
+              "cursor-pointer"
             )}
           >
             {tab.label}
@@ -73,7 +60,7 @@ export default function PurchaseModeTabs({ value, onChange }: PurchaseModeTabsPr
         ))}
       </TabsList>
 
-      {TABS.map((tab) => (
+      {PURCHASE_MODE_TABS.map((tab) => (
         <TabsContent
           key={tab.value}
           value={tab.value}
