+++ src/components/common/tabs/index.tsx
@@ -1,0 +1,1 @@
+export { default as PurchaseModeTabs } from "./PurchaseTabs";
