+++ src/components/common/tabs/tabsConfig.ts
@@ -1,0 +1,4 @@
+export const PURCHASE_MODE_TABS = [
+  { label: "일반 구매", value: "normal", content: "일반 구매 내용입니다." },
+  { label: "자투리 구매", value: "fragment", content: "자투리 구매 내용입니다." },
+] as const;
